requirements.txt
streamlit
duckdb
pandas
plotly
google-generativeai
python-dotenv
sqlalchemy
